+++ examples/next/pages/ui/components/storage/file-uploader/accept-all-file-types/index.page.tsx
@@ -1,0 +1,17 @@
+import { Amplify } from 'aws-amplify';
+import { FileUploader } from '@aws-amplify/ui-react-storage';
+import '@aws-amplify/ui-react/styles.css';
+import amplifyOutputs from './amplify_outputs';
+Amplify.configure(amplifyOutputs);
+
+export function FileUploaderExample() {
+  return (
+    <FileUploader
+      acceptedFileTypes={['*']}
+      path="public/"
+      maxFileCount={1}
+      showThumbnails
+    />
+  );
+}
+export default FileUploaderExample;