+++ packages/react-ai/src/components/AIConversation/views/default/MessageList.tsx
@@ -5,20 +5,22 @@ import {
   AvatarsContext,
   MessageVariantContext,
   RoleContext,
+  useConversationDisplayText,
 } from '../../context';
 import { ConversationMessage } from '../../../../types';
-import { formatDate } from '../../utils';
 import { ControlsContextProps } from '../../context/ControlsContext';
 import {
   ComponentClassName,
   classNameModifier,
   classNames,
 } from '@aws-amplify/ui';
+import { LoadingContext } from '../../context/LoadingContext';
 
 const MessageMeta = ({ message }: { message: ConversationMessage }) => {
   // need to pass this in as props in order for it to be overridable
   const avatars = React.useContext(AvatarsContext);
   const role = React.useContext(RoleContext);
+  const { getMessageTimestampText } = useConversationDisplayText();
   // maybe rename 'avatar' to something else
   const avatar = role === 'assistant' ? avatars?.ai : avatars?.user;
   return (
@@ -27,8 +29,37 @@ const MessageMeta = ({ message }: { message: ConversationMessage }) => {
         {avatar?.username}
       </Text>
       <Text className={ComponentClassName.AIConversationMessageSenderTimestamp}>
-        {formatDate(new Date(message.createdAt))}
+        {getMessageTimestampText(new Date(message.createdAt))}
       </Text>
+    </View>
+  );
+};
+
+const LoadingMessage = () => {
+  const avatars = React.useContext(AvatarsContext);
+  const variant = React.useContext(MessageVariantContext);
+  const avatar = avatars?.ai;
+
+  return (
+    <View
+      className={classNames(
+        ComponentClassName.AIConversationMessage,
+        classNameModifier(ComponentClassName.AIConversationMessage, variant),
+        classNameModifier(ComponentClassName.AIConversationMessage, 'assistant')
+      )}
+    >
+      <View className={ComponentClassName.AIConversationMessageAvatar}>
+        <Avatar isLoading>{avatar?.avatar}</Avatar>
+      </View>
+      <View className={ComponentClassName.AIConversationMessageBody}>
+        <View className={ComponentClassName.AIConversationMessageSender}>
+          <Text
+            className={ComponentClassName.AIConversationMessageSenderUsername}
+          >
+            {avatar?.username}
+          </Text>
+        </View>
+      </View>
     </View>
   );
 };
@@ -67,11 +98,13 @@ const Message = ({ message }: { message: ConversationMessage }) => {
 export const MessageList: ControlsContextProps['MessageList'] = ({
   messages,
 }) => {
+  const isLoading = React.useContext(LoadingContext);
   return (
     <View className={ComponentClassName.AIConversationMessageList}>
       {messages.map((message, i) => (
         <Message key={`message-${i}`} message={message} />
       ))}
+      {isLoading ? <LoadingMessage /> : null}
     </View>
   );
 };