+++ packages/react-ai/CHANGELOG.md
@@ -1,5 +1,18 @@
 # @aws-amplify/ui-react-ai
 
+## 0.1.1
+
+### Patch Changes
+
+- [#5774](https://github.com/aws-amplify/amplify-ui/pull/5774) [`42e2d5d1d`](https://github.com/aws-amplify/amplify-ui/commit/42e2d5d1d22bea316fbefa00ba059bdb7e8e428e) Thanks [@thaddmt](https://github.com/thaddmt)! - fix(ai): remove unnecessary use of Buffer
+
+- [#5782](https://github.com/aws-amplify/amplify-ui/pull/5782) [`c485f1048`](https://github.com/aws-amplify/amplify-ui/commit/c485f1048f158914aaa012e588f9e1caa8bfaa06) Thanks [@thaddmt](https://github.com/thaddmt)! - chore(ai): add timestamp function to displayText
+
+- Updated dependencies [[`9d96dd1fe`](https://github.com/aws-amplify/amplify-ui/commit/9d96dd1fe51212e8e55b0dde816122a6f5672762)]:
+  - @aws-amplify/ui@6.4.1
+  - @aws-amplify/ui-react@6.3.1
+  - @aws-amplify/ui-react-core@3.0.22
+
 ## 0.1.0
 
 ### Minor Changes