+++ packages/react-auth/CHANGELOG.md
@@ -1,5 +1,15 @@
 # @aws-amplify/ui-react-auth
 
+## 0.0.29
+
+### Patch Changes
+
+- Updated dependencies [[`9d96dd1fe`](https://github.com/aws-amplify/amplify-ui/commit/9d96dd1fe51212e8e55b0dde816122a6f5672762)]:
+  - @aws-amplify/ui@6.4.1
+  - @aws-amplify/ui-react@6.3.1
+  - @aws-amplify/ui-react-core@3.0.22
+  - @aws-amplify/ui-react-core-auth@0.0.24
+
 ## 0.0.28
 
 ### Patch Changes