+++ packages/react-native-auth/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @aws-amplify/ui-react-native-auth
 
+## 0.0.27
+
+### Patch Changes
+
+- Updated dependencies [[`9d96dd1fe`](https://github.com/aws-amplify/amplify-ui/commit/9d96dd1fe51212e8e55b0dde816122a6f5672762)]:
+  - @aws-amplify/ui@6.4.1
+  - @aws-amplify/ui-react-core-auth@0.0.24
+  - @aws-amplify/ui-react-native@2.2.9
+
 ## 0.0.26
 
 ### Patch Changes